+++ src/main/java/com/github/warrentode/todecoins/datagen/tags/ItemTagsGen.java
@@ -1411,6 +1411,7 @@ public class ItemTagsGen extends ItemTagsProvider {
                 .addTag(ModTags.Items.CURRENCY_PAPER);
         tag(ForgeTags.Items.TEXTILES)
                 .addTag(ForgeTags.Items.BARK)
+                .addTag(ForgeTags.Items.LEATHER)
                 .addTag(ForgeTags.Items.FIBRE)
                 .addTag(ForgeTags.Items.STRING)
                 .addTag(ForgeTags.Items.PAPER)
@@ -1418,6 +1419,7 @@ public class ItemTagsGen extends ItemTagsProvider {
                 .addTag(ForgeTags.Items.THREAD)
                 .addTag(ForgeTags.Items.CURRENCY_PAPER)
                 .addTag(ForgeTags.Items.CURRENCY_FIBER);
+        tag(ForgeTags.Items.LEATHER);
         tag(ForgeTags.Items.BARK);
         tag(ForgeTags.Items.THREAD);
         tag(ForgeTags.Items.CURRENCY_PAPER);